ns-3.38 release

+++ src/lte/test/lte-test-fdbet-ff-mac-scheduler.cc
@@ -194,6 +194,10 @@ LenaTestFdBetFfMacSchedulerSuite::LenaTestFdBetFfMacSchedulerSuite()
         TestCase::QUICK);
 }
 
+/**
+ * \ingroup lte-test
+ * Static variable for test initialization
+ */
 static LenaTestFdBetFfMacSchedulerSuite lenaTestFdBetFfMacSchedulerSuite;
 
 // --------------- T E S T - C A S E   # 1 ------------------------------
@@ -519,7 +523,7 @@ LenaFdBetFfMacSchedulerTestCase2::DoRun()
      */
     for (int i = 0; i < m_nUser; i++)
     {
-        double thrRatio = (double)1 / m_nUser;
+        double thrRatio = 1.0 / m_nUser;
         double estThrRatio = (double)dlDataRxed.at(i) / totalData;
         NS_LOG_INFO("\tUser " << i << " thrRatio " << thrRatio << " estThrRatio " << estThrRatio);
         NS_TEST_ASSERT_MSG_EQ_TOL(estThrRatio, thrRatio, tolerance, " Unfair Throughput!");
